refactor: WordMeaningBlock as a separated component
